Reworked FirebaseConfig to match rules in DB
Reworked FirebaseConfig CreateUser to pull the uid from the auth entry and create a new user entry with that uid as the id. Our rules now give us protection and only allow our users who are admins to create events. And only allow the user themselves to update their profile

-Adjusted sign out method to fit in FirebaseConfig.ts
-Added documentation to changes
-Reworked onAuthStateChanged
-Removed unnecessary debugging lines